Centering highlight on IconButton
